+++ src/main/java/org/synyx/urlaubsverwaltung/availability/api/FreeTimeAbsenceProvider.java
@@ -70,7 +70,7 @@ class FreeTimeAbsenceProvider extends AbstractTimedAbsenceProvider {
                 currentDay);
 
         if (!workingTimeOrNot.isPresent()) {
-            throw new IllegalStateException("Person " + person + " does not have workingTime configured");
+            throw new FreeTimeAbsenceException("Person " + person + " does not have workingTime configured");
         }
 
         WorkingTime workingTime = workingTimeOrNot.get();